COMAAS-280: Register Comaas with Consul
This used to be done by /etc/init.d/mp-replyts but that isn't called by
deploy.py, which instead calls daemontools directly. This is part of the
fix for the 502s issue that MP was having.

Let it be known that it is no longer possible to configure the http port
that Comaas is running on using the property 'replyts.http.port'.
Instead, if you want to run on anything but the default port 8080, you
will have to set an environment variable: COMAAS_HTTP_PORT. This was
added earlier already, as part of the preparations for running under
Nomad.
I understand that this means that we cannot use the QA environment as it
currently is anymore, since that requires multiple instances of Comaas
to run on the same host on different ports. QA is currently not being
deployed to, and we should let Nomad find an open port for a new service
instead of hard coding it.
I had to remove 'replyts.http.port' from the properties because of the
way we are loading properties from the various sources. In
CloudDiscoveryConfiguration.initializeDiscovery() we kick off the
application using 'lifecycle.onApplicationEvent(...)' which loads the
various property sources as well. In this call we initialize a
EmbeddedServletContainerInitializedEvent with an overridden getPort()
method. The port must thus be configured already... but this happens
before loading the property sources! Chicken & egg.

This patch needs an accompanying change in Salt that sets a new
environment variable called COMAAS_FABIO_URL_PREFIXES. This should hold
the urlprefixes that are sent to Consul as tags and are then picked up
by Fabio.

As a bonus, I've added hashi-ui as the default Consul web viewer when
running with Docker Compose.

Reviewed-on: https://gerrit.ecg.so/60700
Tested-by: Gerrit-At-Gitmaster001
Tested-by: Comaas Jenkins <DL-eCG-comaas-dev@ebay.com>

+++ integration-tests/core-integration-test-api/src/main/java/com/ecg/replyts/integration/test/ReplytsRunner.java
@@ -4,6 +4,8 @@ import com.ecg.replyts.core.runtime.ReplyTS;
 import com.ecg.replyts.integration.cassandra.CassandraIntegrationTestProvisioner;
 import com.google.common.io.Files;
 import org.elasticsearch.client.Client;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.xml.XmlBeanDefinitionReader;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.springframework.core.env.PropertiesPropertySource;
@@ -14,15 +16,14 @@ import org.subethamail.wiser.Wiser;
 import org.subethamail.wiser.WiserMessage;
 
 import java.io.File;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Properties;
-import java.util.UUID;
+import java.util.*;
 
+import static com.ecg.replyts.integration.test.support.IntegrationTestUtils.setEnv;
 import static com.google.common.base.Preconditions.checkNotNull;
 
 public final class ReplytsRunner {
+    private static final Logger LOG = LoggerFactory.getLogger(ReplytsRunner.class);
+
     public static final String DEFAULT_CONFIG_RESOURCE_DIRECTORY = "/integrationtest-conf";
 
     private static final String ELASTIC_SEARCH_PREFIX = "comaas_integration_test_";
@@ -30,14 +31,13 @@ public final class ReplytsRunner {
     private final File dropFolder = Files.createTempDir();
     private final Integer httpPort = OpenPortFinder.findFreePort();
     private final Integer smtpOutPort = OpenPortFinder.findFreePort();
-    private final String elasticClusterName = ELASTIC_SEARCH_PREFIX + UUID.randomUUID();
 
     private Wiser wiser = new Wiser(smtpOutPort);
     private AnnotationConfigApplicationContext context;
     private Client searchClient;
     private MailInterceptor mailInterceptor;
 
-    public ReplytsRunner(Properties testProperties, String configResourcePrefix, Class<?>[] configurations) {
+    ReplytsRunner(Properties testProperties, String configResourcePrefix, Class<?>[] configurations) {
         wiser.start();
 
         if (!wiser.getServer().isRunning())
@@ -60,12 +60,13 @@ public final class ReplytsRunner {
 
             properties.put("persistence.cassandra.core.endpoint", CassandraIntegrationTestProvisioner.getEndPoint());
             properties.put("persistence.cassandra.mb.endpoint", CassandraIntegrationTestProvisioner.getEndPoint());
-            properties.put("replyts.http.port", String.valueOf(httpPort));
+            setEnv("COMAAS_HTTP_PORT", httpPort.toString());
             properties.put("replyts.ssl.enabled", "false");
             properties.put("delivery.smtp.port", String.valueOf(smtpOutPort));
 
             properties.put("mailreceiver.filesystem.dropfolder", dropFolder.getAbsolutePath());
 
+            String elasticClusterName = ELASTIC_SEARCH_PREFIX + UUID.randomUUID();
             properties.put("search.es.clustername", elasticClusterName);
 
             properties.put("node.passive", "true");
@@ -122,7 +123,7 @@ public final class ReplytsRunner {
         return httpPort;
     }
 
-    public Client getSearchClient() {
+    Client getSearchClient() {
         if (searchClient == null) {
             searchClient = context.getBean(Client.class);
 
@@ -134,23 +135,29 @@ public final class ReplytsRunner {
         return searchClient;
     }
 
-    public File getDropFolder() {
+    File getDropFolder() {
         return dropFolder;
     }
 
     private void deleteDirectory(File directory) {
         File[] files = directory.listFiles();
-        for (File file : files) {
-            if (file.isDirectory()) {
-                deleteDirectory(file);
-            } else {
-                file.delete();
+        if (files != null) {
+            for (File file : files) {
+                if (file.isDirectory()) {
+                    deleteDirectory(file);
+                } else {
+                    if (!file.delete()) {
+                        LOG.warn("Could not delete file {}", file);
+                    }
+                }
             }
         }
-        directory.delete();
+        if (!directory.delete()) {
+            LOG.warn("Could not delete directory {}", directory);
+        }
     }
 
-    public MailInterceptor getMailInterceptor() {
+    MailInterceptor getMailInterceptor() {
         return mailInterceptor;
     }
 }
